Publish zed wasm (#4)
This PR adds a github workflow that is triggered whenever a package.json file changes in one of the packages folders. It then checks if the version in package.json is different from the latest version on npm. If so, it publishes it!

We have a matrix setup to run this workflow for all publishable packages. At the moment, that is zed-js, zed-node, and zed-wasm.

I also had to re-format a whole bunch of files to remove the ".js" extension.